Create bg_update.c
