+++ ldscript.txt
@@ -35,6 +35,7 @@ SECTIONS
         build/sa2/src/task.o(.text);
         build/sa2/src/sprite_1.o(.text);
         build/sa2/asm/sprite.o(.text);
+        build/sa2/src/bg_update.o(.text);
         build/sa2/src/animation_commands_extra.o(.text);
         build/sa2/asm/sprite_2.o(.text);
         build/sa2/src/sprite_3.o(.text);
